feat: Dynamic Sitemap Generation
Following [this
guide](https://nextjs.org/docs/app/api-reference/file-conventions/metadata/sitemap),
I implemented the new sitemap generator.

I chose to hard-code the domain in the file if we need to change it at
any time. I'm not really sure of any other way to fetch the URL as it
isn't stored in any other files.

The current problem with the implementation is that the sitemap is
regenerated on each page load, thus:
- Sitemap loads more slowly (reruns on each request)
- The `lastModified` date becomes whatever the date is of page load