feat: Add Github and Discord to homepage
I added the links to the card following the code in `app/[slug]/page.tsx`.

I used the same icon repository (ant-icons) to choose the icons. The
Github icon is self-explanatory, while the Message icon was chosen
because there isn't a Discord icon in the icon repository.

Icons were chosen over the full URL as the default footer styling makes
it an ugly two-column layout and has text-wrapping. As such, I
just opted to use the icons (and made them large) to save the headache of restyling.