fix: Better workaround for theme rendering
Work smarter, not harder:tm:

Compared to the old strategy of having a state hook which tells the page
when to render and when not to, it now abuses the fact that this is a
Tailwind CSS website, which means we can use Tailwind CSS classes.

The "invisible" class sets the visibility to hidden, causing the element
to not render. As such, the default theme does not render the page.

Once the theme is changed when reading the local storage, the page will
render with the new theme selected, causing no where ghosting of the UI
when the page changes themes.

There's still the flicker though :/