fix: Fix infinite API polling
For some reason, the component used to keep requesting the data from the
API even though it has already responded. This happened mainly in
production mode, though occasionally happened in development.

The fix that I found was to make the component wait until at least react
as loaded on the client for it to finally render. I suspect the main
problem was that the site would be trying to re-render itself beween the
time to request the API and the time between it getting it. I think it
then flooded the thread, blocking the JS from running. Probably wasn't
helping that NEXTUI is completely client-side and the site only renders
when react has loaded, so sending hundreds of fetch requests quickly
probably did it in.

There was also a weird bug where the API kept double-requesting the API,
which seems to have been fixed. (Probably due to a request
pre-and-after darkmode switch)

God I love just trying random stuff. Works surprisingly well.