refactor: Let EventList take prop to set group
Now you can pass the string (or rather, url) into the function and get
the associated calendar. This is the same way as the slug currently
works as it just parses the URL and loads the accompanying data.

Also added the emoji thing to make it easier to see what I'm doing...