refactor: Switch to RestAPI (& 3 weeks)
Switched it to 3 weeks as requested by Jorge but I (incorrectly)
assumed he meant one week before and one week after so that's what I
did ;).

I even made a little graphic for it.

In other news, the reason there isn't a BEARER thing is because that
causes errors. Yes, the google docs tell you to use it but I guess
that's for oauth only and they just don't say it. :)