refactor: Set it to correct 3 weeks
Originally, I did it one week before and one week after. I thought it
would be fine but when I rolled out alphabetical sort for the
calendar, I realized it was pretty stupid. I want newest events to be
at the top, and it's dumb to include events that happened a week ago.

Considering not even concluding the whole current week.