feat: End of event list card
There's now a card that says when you're at the bottom of the events
list. Originally this was only going to exist for pages with no
events (like a "no events found, check in later"). Realized this would
just be better (and less cluttery) to just have it for all of the lists.