refactor: Move EventList wrapping card to uses
I moved the EventList parent cards to the places that it's used to
allow for more variance in it's width and the like.

Still doesn't really fix the problem that the styling in /calendar, as
it's not possible to style the interior CardBody element (which would
allow for it to be styled to do the flex stuff & card size stuff).

At least it's slightly cleaner.